Fixed querying the GraphQL server.
Lots of changes here. Got carried away. The resolvers in the GraphQL
server for "allDatasets" and "datasets" didn't work. So I fixed them.
Also fixed the querying of the GraphQL server. Currently about to switch
from GraphQL-Yoga to Apollo (also adding the Apollo Client) so that
running the server from the container can work too. The ports are being
published, but the IP address listened on is incorrect and Yoga doesn't
allow me to change that. Sick...