Fixed GraphQL server containerisation.
Making requests to the container running the GraphQL server works. I
just had some problems with IP addresses is all. Everything's fine now
though. Switched from GraphQL Yoga to Apollo too. I'm going to add its
client then finish working on the resolver for the data for the File
type and that should be it for the GraphQL server for now.