It works chief.
Well the GraphQL server that is. A JSON scalar was added so I can now
send JSON objects as the type of a field in the selection set. Now I
just need to work on the schema and parameterising things like the file
name to retrieve, then I can make the REST API a proxy to the GraphQL
server. Nice :D