Cleaned up the data route. RIP choochoo.
The last commit about choochoo didn't actually get rid of it. But this
one changes that. The console log statement known as "choochoo" is no
more. The path with the same name has also left us. Tis a sad day. Oh, I
also converted the iris data to a JSON file too.